+++ R/epinow.R
@@ -133,14 +133,6 @@ epinow <- function(data,
   assert_string(id)
   assert_logical(verbose)
 
-  if (is.null(CrIs) || length(CrIs) == 0 || !is.numeric(CrIs)) {
-    futile.logger::flog.fatal(
-      "At least one credible interval must be specified",
-      name = "EpiNow2.epinow"
-    )
-    stop("At least one credible interval must be specified")
-  }
-
   if (is.null(forecast)) {
     forecast <- forecast_opts(horizon = 0)
   }
@@ -212,11 +204,6 @@ epinow <- function(data,
       id = id
     )
 
-    if (!output["fit"]) {
-      estimates$fit <- NULL
-      estimates$args <- NULL
-    }
-
     save_estimate_infections(estimates, target_folder,
       samples = output["samples"],
       return_fit = output["fit"]
@@ -230,7 +217,7 @@ epinow <- function(data,
     )
 
     # report estimates --------------------------------------------------------
-    summary <- summary.estimate_infections(estimates,
+    summary <- summary(estimates,
       return_numeric = TRUE,
       target_folder = target_folder,
       CrIs = CrIs
@@ -257,6 +244,11 @@ epinow <- function(data,
         summary,
         samples = output["samples"]
       )
+      if (output["fit"]) {
+        out$estimates$fit <- estimates$fit
+        out$estimates$args <- estimates$args
+      }
+      out$estimates$observations <- estimates$observations
       return(out)
     } else {
       return(invisible(NULL))